+++ src/js/overrides.js
@@ -2,7 +2,7 @@
 
                   This file is part of the Buildbotics firmware.
 
-         Copyright (c) 2015 - 2021, Buildbotics LLC, All rights reserved.
+         Copyright (c) 2015 - 2022, Buildbotics LLC, All rights reserved.
 
           This Source describes Open Hardware and is licensed under the
                                   CERN-OHL-S v2.
@@ -27,47 +27,33 @@
 
 'use strict'
 
+var api = require('./api');
+
 
 module.exports = {
-  template: "#io-pins-template",
-  props: ['template', 'state'],
+  template: '#overrides-template',
+  props: ['state', 'template'],
 
-
-  components: {
-    'io-pins-row': {
-      template: "#io-pins-row-template",
-      props: ['pin'],
-      replace: true
+  data() {
+    return {
+      speed: 1,
+      feed: 1
     }
   },
 
 
-  computed: {
-    columns: () => ['pin', 'state', 'function', '', 'pin', 'state', 'function'],
-    rows: function () {return Math.ceil(this.io_pins.length / 2.0)},
+  ready() {
+    this.speed = this.state.so || 1
+    this.feed  = this.state.fo || 1
+  },
 
 
-    io_pins: function () {
-      let l = []
-      let io_map = this.template['io-map']
-      let pins = io_map.pins
-      let modes = io_map.template.mode.values
-      let functions = io_map.template.function.values
+  methods: {
+    override_feed: function () {api.put('override/feed/' + this.feed)},
 
-      for (let i = 0; i < pins.length; i++) {
-        let c     = String.fromCharCode(97 + i)
-        let func  = functions[this.state[c + 'io']]
-        let state = func == 'disabled' ? 'disabled' : this.state[c + 'is']
-        let type  = pins[i].type
-        let mode  = ''
 
-        if (func != 'disabled' && type != 'analog')
-          mode = modes[this.state[c + 'im']]
-
-        l.push({id: pins[i].id, state, mode, func, type})
-      }
-
-      return l
+    override_speed: function () {
+      api.put('override/speed/' + this.speed)
     }
   }
 }